third_party/WebKit/LayoutTests/inspector/sass/test-mapping-good-expected.txt

Verify WI.ASTSourceMap rebaselines.

Mapped CSS: 14
Mapped SCSS: 14
No stale entries found.

Running: testRebase
Mapped CSS: 14
Mapped SCSS: 14
No stale entries found.
